refactor: use ts exts for relative paths
