+++ src/cli.ts
@@ -28,11 +28,12 @@ import {
   path,
   stdin,
   VERSION,
-} from './index.js'
-import { installDeps, parseDeps } from './deps.js'
-import { startRepl } from './repl.js'
-import { randomId, bufToString } from './util.js'
-import { createRequire, type minimist } from './vendor.js'
+} from './index.ts'
+import { installDeps, parseDeps } from './deps.ts'
+import { startRepl } from './repl.ts'
+import { randomId, bufToString } from './util.ts'
+import { transformMarkdown } from './md.ts'
+import { createRequire, type minimist } from './vendor.ts'
 
 const EXT = '.mjs'
 const EXT_RE = /^\.[mc]?[jt]sx?$/
@@ -206,84 +207,13 @@ async function readScriptFromHttp(remote: string): Promise<string> {
   return res.text()
 }
 
+export { transformMarkdown }
+
 export function injectGlobalRequire(origin: string): void {
   const __filename = path.resolve(origin)
   const __dirname = path.dirname(__filename)
   const require = createRequire(origin)
   Object.assign(globalThis, { __filename, __dirname, require })
-}
-
-export function transformMarkdown(buf: Buffer | string): string {
-  const output = []
-  const tabRe = /^(  +|\t)/
-  const codeBlockRe =
-    /^(?<fence>(`{3,20}|~{3,20}))(?:(?<js>(js|javascript|ts|typescript))|(?<bash>(sh|shell|bash))|.*)$/
-  let state = 'root'
-  let codeBlockEnd = ''
-  let prevLineIsEmpty = true
-  for (const line of bufToString(buf).split(/\r?\n/)) {
-    switch (state) {
-      case 'root':
-        if (tabRe.test(line) && prevLineIsEmpty) {
-          output.push(line)
-          state = 'tab'
-          continue
-        }
-        const { fence, js, bash } = line.match(codeBlockRe)?.groups || {}
-        if (!fence) {
-          prevLineIsEmpty = line === ''
-          output.push('// ' + line)
-          continue
-        }
-        codeBlockEnd = fence
-        if (js) {
-          state = 'js'
-          output.push('')
-        } else if (bash) {
-          state = 'bash'
-          output.push('await $`')
-        } else {
-          state = 'other'
-          output.push('')
-        }
-        break
-      case 'tab':
-        if (line === '') {
-          output.push('')
-        } else if (tabRe.test(line)) {
-          output.push(line)
-        } else {
-          output.push('// ' + line)
-          state = 'root'
-        }
-        break
-      case 'js':
-        if (line === codeBlockEnd) {
-          output.push('')
-          state = 'root'
-        } else {
-          output.push(line)
-        }
-        break
-      case 'bash':
-        if (line === codeBlockEnd) {
-          output.push('`')
-          state = 'root'
-        } else {
-          output.push(line)
-        }
-        break
-      case 'other':
-        if (line === codeBlockEnd) {
-          output.push('')
-          state = 'root'
-        } else {
-          output.push('// ' + line)
-        }
-        break
-    }
-  }
-  return output.join('\n')
 }
 
 export function isMain(
