+++ src/globals.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import * as _ from './index.js'
+import * as _ from './index.ts'
 
 Object.assign(globalThis, _)
 // TODO: global types not working with jsr.io
