+++ tsconfig.json
@@ -8,6 +8,7 @@
     "outDir": "./build",
     "declaration": true,
     "emitDeclarationOnly": true,
+    "allowImportingTsExtensions": true,
     "types": []
   },
   "include": ["./src/**/*"],
